Extract assessmentStatusUtils
As we will later want to use statusName() in our integration tests, we extract this and related functions into their own file. This will allow us to use them in our integration tests without indirectly pulling in restUtils, which fails to compile in the context of integration tests
